rename: 바른 이름의 Fragment로 변경 하고, 바른 위치의 package로 이동 하였습니다.
다른 분이 초기에 임시로 만들어주신 담당 프래그먼트도 지웠습니다
