+++ app/src/main/res/layout/fragment_my_video.xml
@@ -5,7 +5,7 @@
     android:id="@+id/cl"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyVideoFragment">
+    tools:context=".presentation.myvideo.MyVideoFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_my_video"
